Add Bean Validation

+++ src/main/java/com/jpm/commerce/services/ProductService.java
@@ -3,7 +3,11 @@ package com.jpm.commerce.services;
 import com.jpm.commerce.dto.ProductDTO;
 import com.jpm.commerce.entities.Product;
 import com.jpm.commerce.repositories.ProductRepository;
+import com.jpm.commerce.services.exceptions.DatabaseException;
+import com.jpm.commerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -16,7 +20,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
         //Optional<Product> result= repository.findById(id);
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(()->new ResourceNotFoundException("Recurso nao encontrad"));
        // ProductDTO dto = new ProductDTO(product.getId(), product.getName(), product.getDescription(), product.getPrice(), product.getImgUrl());
         return new ProductDTO(product);
     }
@@ -35,15 +39,29 @@ public class ProductService {
     }
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try{
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
 
     }
     @Transactional
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw  new ResourceNotFoundException("Recurso nao encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha na integriidade referencial");
+        }
+
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
